Make events not show between sprints without a gap
Previously, when sprints were back-to-back (one day difference in dates),
the events would show between them on the project details page. Now, they
only show if the next sprint doesn't start on the day after the previous
one finishes.

Added a function to DateUtils and a way to access it from Thymeleaf.